requirements.txt
bitmex
websockets
bitmex_websocket
mysql-connector